// src/app/models/ReporteCierre.ts
export class ReporteCierre{
   

    //RECIBO MASIVO

    idEmpresa: number | null = null;
	idSede: number | null = null;
	idCiclo: number | null = null;
	idSucursal: number | null = null;
	idSector: number | null = null;
	anio: string | null = null;
	mes: number | null = null;
	usuarioCreacion: string | null = null;


	fecDesde: string | null = null;
	fecDesdeDpl: string | null = null;
	fecHasta: string | null = null;
    fecHastaDpl: string | null = null;
	idTipo: string | null = null;
	



    
}